Changed encoderTicks to int.

+++ src/opcontrol.c
@@ -153,10 +153,10 @@ float pidController(float Kp, float Ki, float Kd, float setpoint,
  * critical to make sure that the derivative doesn't add too much noise.
  */
 void updateShooterSpeedTask(void *ignore) {
-	float encoderTicks = encoderGet(shooterEncoder);
+	int encoderTicks = encoderGet(shooterEncoder);
 	//shooterSpeed = ((encoderTicks / 360.0) / (1 / (SHOOTER_ENCODER_POLL_RATE) * 60));
 	//encoderReset(shooterEncoder);
-	printf("Shooter speed: %f \n", encoderTicks);
+	printf("Shooter speed: %d \n", encoderTicks);
 	taskDelayUntil(&lastShooterSpeedLoopStopTime,
 			(1 / SHOOTER_ENCODER_POLL_RATE) * 1000);
 }
